Add doc, move test methods into test modules
