+++ datafusion/optimizer/src/decorrelate_where_exists.rs
@@ -199,8 +199,6 @@ impl SubqueryInfo {
 mod tests {
     use super::*;
     use crate::test::*;
-    use crate::utils::assert_optimized_plan_eq;
-    use crate::utils::assert_optimizer_err;
     use datafusion_common::Result;
     use datafusion_expr::{
         col, exists, lit, logical_plan::LogicalPlanBuilder, not_exists,
